.history/backend/requirements_20250517164634.txt
fastapi
uvicorn
tortoise-orm
aerich
aiosqlite
python-jose[cryptography]
passlib[bcrypt]
python-multipart 